[POSUI-246] fix the issue that \1, \2, and \3 escape characters were not working in the text
