+++ app/src/main/java/com/paxus/pay/poslinkui/demo/entry/poslink/ShowMessageFragment.java
@@ -157,7 +157,7 @@ public class ShowMessageFragment extends BaseEntryFragment {
                 wrapper.setIndex(index);
 
                 JSONObject msgObj = obj.getJSONObject("MsgInfo");
-                wrapper.setMsgInfo(new MsgInfo(msgObj.getString("msg1"), msgObj.getString("msg2")));
+                wrapper.setMsgInfo(new MsgInfoWrapper.MsgInfo(msgObj.getString("msg1"), msgObj.getString("msg2")));
 
                 list.add(wrapper);
             }
@@ -170,39 +170,6 @@ public class ShowMessageFragment extends BaseEntryFragment {
         return null;
     }
 
-    //[{"index":null,"MsgInfo":{"msg1":"","msg2":"8654"}},{"index":null,"MsgInfo":{"msg1":"","msg2":"8654"}},{"index":null,"MsgInfo":{"msg1":"","msg2":"8654"}},{"index":null,"MsgInfo":{"msg1":"","msg2":"8654"}},{"index":null,"MsgInfo":{"msg1":"","msg2":"8654"}},{"index":null,"MsgInfo":{"msg1":"976&","msg2":"8654"}},{"index":null,"MsgInfo":{"msg1":"976&","msg2":"8654"}},{"index":null,"MsgInfo":{"msg1":"976&","msg2":"8654"}},{"index":null,"MsgInfo":{"msg1":"976&","msg2":"8654"}},{"index":null,"MsgInfo":{"msg1":"976&","msg2":"8654"}},{"index":null,"MsgInfo":{"msg1":"976&","msg2":"8654"}}]
-    public static class MsgInfoWrapper {
-        private String index;
-        private MsgInfo msgInfo;
-
-        public MsgInfoWrapper() {
-        }
-
-        public String getIndex() {
-            return index;
-        }
-
-        public void setIndex(String index) {
-            this.index = index;
-        }
-
-        public MsgInfo getMsgInfo() {
-            return msgInfo;
-        }
-
-        public void setMsgInfo(MsgInfo msgInfo) {
-            this.msgInfo = msgInfo;
-        }
-    }
-
-    public static class MsgInfo{
-        public String msg1;
-        public String msg2;
-        public MsgInfo(String msg1,String msg2){
-            this.msg1 = msg1;
-            this.msg2 = msg2;
-        }
-    }
 
     @Override
     protected TextField[] focusableTextFields() {
